[Refactor] #40 - TodayWeeklyCalendarViewModel 구조 개선
MARK 주석 구분
Date 관련 로직 간소화
불필요한 코드 제거
