+++ Memento-iOS/Memento-iOS/Source/Extension/Date+.swift
@@ -12,6 +12,10 @@ extension Date {
         return Calendar.current.startOfDay(for: self)
     }
     
+    var endOfDay: Date {
+        return Calendar.current.date(byAdding: DateComponents(day: 1, second: -1), to: startOfDay)!
+    }
+    
     /// 특정 포맷의 날짜 문자열 반환
     func formattedDate(with format: String, timeZone: TimeZone = .current) -> String {
         let dateFormatter = DateFormatter()
